+++ gesturefx/src/main/java/net/kurobako/gesturefx/NamedAffine.java
@@ -6,6 +6,7 @@ import javafx.beans.property.ReadOnlyDoubleProperty;
  * A simple abstraction over some affine transformation. Some properties are intentionally
  * left out as those are not used in {@link GesturePane}
  */
+@SuppressWarnings("unused")
 public interface NamedAffine {
 	/**
 	 * @return the scale on the X axis